+++ packages/core/src/agent/agent-network.test.ts
@@ -3416,3 +3416,609 @@ describe('Agent - network - finalResult real-world scenarios', () => {
     expect(finishEvent.payload.result).toContain('2 iterations');
   });
 });
+
+describe('Agent - network - structured output', () => {
+  const memory = new MockMemory();
+  const requestContext = new RequestContext();
+
+  // Schema for structured network output
+  const resultSchema = z.object({
+    summary: z.string().describe('A brief summary of the task result'),
+    recommendations: z.array(z.string()).describe('List of recommendations'),
+    confidence: z.number().min(0).max(1).describe('Confidence score'),
+  });
+
+  // Expected structured result
+  const structuredResult = {
+    summary: 'Task completed successfully',
+    recommendations: ['Recommendation 1', 'Recommendation 2'],
+    confidence: 0.95,
+  };
+
+  // Mock scorer that always passes
+  const alwaysPassScorer = {
+    id: 'always-pass',
+    name: 'Always Pass Scorer',
+    run: vi.fn().mockResolvedValue({ score: 1, reason: 'Always passes' }),
+  };
+
+  /**
+   * Creates a mock model that handles the network flow:
+   * 1. First call: Routing decision
+   * 2. Second call: Structured output generation (when schema is provided)
+   */
+  function createNetworkMockModel(options: { routingResponse?: object; structuredResult?: object }) {
+    const routingResponse = JSON.stringify(
+      options.routingResponse ?? {
+        primitiveId: 'none',
+        primitiveType: 'none',
+        prompt: '',
+        selectionReason: 'Task handled directly',
+      },
+    );
+
+    const structuredResultJson = JSON.stringify(options.structuredResult ?? structuredResult);
+
+    let callCount = 0;
+
+    return new MockLanguageModelV2({
+      doGenerate: async () => {
+        callCount++;
+        // First call is routing, subsequent calls return structured result
+        const text = callCount === 1 ? routingResponse : structuredResultJson;
+        return {
+          rawCall: { rawPrompt: null, rawSettings: {} },
+          finishReason: 'stop',
+          usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
+          content: [{ type: 'text', text }],
+          warnings: [],
+        };
+      },
+      doStream: async () => {
+        callCount++;
+        // First call is routing, subsequent calls return structured result
+        const text = callCount === 1 ? routingResponse : structuredResultJson;
+
+        // Stream JSON - for non-routing calls, stream in chunks
+        const chunks = [
+          { type: 'stream-start' as const, warnings: [] },
+          {
+            type: 'response-metadata' as const,
+            id: `id-${callCount}`,
+            modelId: 'mock-model-id',
+            timestamp: new Date(0),
+          },
+          { type: 'text-start' as const, id: `text-${callCount}` },
+          { type: 'text-delta' as const, id: `text-${callCount}`, delta: text },
+          { type: 'text-end' as const, id: `text-${callCount}` },
+          {
+            type: 'finish' as const,
+            finishReason: 'stop' as const,
+            usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
+          },
+        ];
+
+        return {
+          stream: convertArrayToReadableStream(chunks),
+          rawCall: { rawPrompt: null, rawSettings: {} },
+          warnings: [],
+        };
+      },
+    });
+  }
+
+  it('should expose .object getter that returns a Promise', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test task', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+    });
+
+    // Consume stream
+    for await (const _chunk of stream) {
+      // Process
+    }
+
+    // Test that .object getter exists and is a Promise
+    expect(stream).toHaveProperty('object');
+    expect(typeof (stream as any).object?.then).toBe('function');
+  });
+
+  it('should expose .objectStream getter that returns a ReadableStream', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test task', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+    });
+
+    // Test that .objectStream getter exists
+    expect(stream).toHaveProperty('objectStream');
+    expect((stream as any).objectStream).toBeInstanceOf(ReadableStream);
+  });
+
+  it('should return typed object when structuredOutput.schema is provided', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test structured output',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Analyze this', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    // Consume stream
+    for await (const _chunk of stream) {
+      // Process
+    }
+
+    // Get the structured object
+    const result = await stream.object;
+
+    // These assertions verify the feature is implemented
+    expect(result).toBeDefined();
+    expect(result).not.toBeUndefined();
+    expect(result!.summary).toBe(structuredResult.summary);
+    expect(result!.recommendations).toEqual(structuredResult.recommendations);
+    expect(result!.confidence).toBe(structuredResult.confidence);
+  });
+
+  it('should emit network-object-result chunk with typed object', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test object chunks',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Check chunks', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    let objectResultChunk: any = null;
+
+    for await (const chunk of stream) {
+      if (chunk.type === 'network-object-result') {
+        objectResultChunk = chunk;
+      }
+    }
+
+    // This will pass when the feature emits network-object-result chunks
+    expect(objectResultChunk).not.toBeNull();
+    expect(objectResultChunk.payload.object).toEqual(structuredResult);
+  });
+
+  it('should include object property in NetworkFinishPayload', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test finish payload',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Check finish payload', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    let finishPayload: any = null;
+
+    for await (const chunk of stream) {
+      if (chunk.type === 'network-execution-event-finish') {
+        finishPayload = chunk.payload;
+      }
+    }
+
+    expect(finishPayload).toBeDefined();
+    // NetworkFinishPayload should include object property
+    expect(finishPayload).toHaveProperty('object');
+    expect(finishPayload.object).toEqual(structuredResult);
+  });
+
+  it('should stream partial objects via objectStream', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test streaming',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Stream test', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    const partialObjects: any[] = [];
+
+    // objectStream should be defined
+    const objectStream = stream.objectStream;
+    expect(objectStream).toBeDefined();
+
+    // Start consuming objectStream in background
+    const objectStreamPromise = (async () => {
+      for await (const partial of objectStream) {
+        partialObjects.push(partial);
+      }
+    })();
+
+    // Consume main stream
+    for await (const _chunk of stream) {
+      // Process
+    }
+
+    // Wait for objectStream to finish
+    await objectStreamPromise;
+
+    // Should have received at least one partial object
+    expect(partialObjects.length).toBeGreaterThan(0);
+  });
+
+  it('should emit network-object chunks during streaming', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test object chunks',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Partial chunks test', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    const objectChunks: any[] = [];
+
+    for await (const chunk of stream) {
+      if (chunk.type === 'network-object') {
+        objectChunks.push(chunk);
+      }
+    }
+
+    // network-object chunks should be emitted during streaming
+    expect(objectChunks.length).toBeGreaterThan(0);
+  });
+
+  it('should generate structured output after sub-agent completes', async () => {
+    const subAgentResponse = 'Detailed analysis results from sub-agent.';
+
+    const subAgentMockModel = new MockLanguageModelV2({
+      doGenerate: async () => ({
+        rawCall: { rawPrompt: null, rawSettings: {} },
+        finishReason: 'stop',
+        usage: { inputTokens: 20, outputTokens: 40, totalTokens: 60 },
+        content: [{ type: 'text', text: subAgentResponse }],
+        warnings: [],
+      }),
+      doStream: async () => ({
+        stream: convertArrayToReadableStream([
+          { type: 'stream-start', warnings: [] },
+          { type: 'response-metadata', id: 'id-0', modelId: 'sub-agent-model', timestamp: new Date(0) },
+          { type: 'text-start', id: 'text-1' },
+          { type: 'text-delta', id: 'text-1', delta: subAgentResponse },
+          { type: 'text-end', id: 'text-1' },
+          {
+            type: 'finish',
+            finishReason: 'stop',
+            usage: { inputTokens: 20, outputTokens: 40, totalTokens: 60 },
+          },
+        ]),
+        rawCall: { rawPrompt: null, rawSettings: {} },
+        warnings: [],
+      }),
+    });
+
+    const subAgent = new Agent({
+      id: 'research-agent',
+      name: 'Research Agent',
+      description: 'Performs detailed research',
+      instructions: 'Research topics thoroughly',
+      model: subAgentMockModel,
+    });
+
+    // Routing agent selects sub-agent first, then handles structured output
+    const routingSelectAgent = JSON.stringify({
+      primitiveId: 'research-agent',
+      primitiveType: 'agent',
+      prompt: 'Analyze this topic',
+      selectionReason: 'Delegating to research agent',
+    });
+
+    const structuredResultJson = JSON.stringify(structuredResult);
+
+    let callCount = 0;
+    const routingMockModel = new MockLanguageModelV2({
+      doGenerate: async () => {
+        callCount++;
+        const text = callCount === 1 ? routingSelectAgent : structuredResultJson;
+        return {
+          rawCall: { rawPrompt: null, rawSettings: {} },
+          finishReason: 'stop',
+          usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
+          content: [{ type: 'text', text }],
+          warnings: [],
+        };
+      },
+      doStream: async () => {
+        callCount++;
+        const text = callCount === 1 ? routingSelectAgent : structuredResultJson;
+        return {
+          stream: convertArrayToReadableStream([
+            { type: 'stream-start', warnings: [] },
+            { type: 'response-metadata', id: 'id-0', modelId: 'routing-model', timestamp: new Date(0) },
+            { type: 'text-start', id: 'text-1' },
+            { type: 'text-delta', id: 'text-1', delta: text },
+            { type: 'text-end', id: 'text-1' },
+            {
+              type: 'finish',
+              finishReason: 'stop',
+              usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
+            },
+          ]),
+          rawCall: { rawPrompt: null, rawSettings: {} },
+          warnings: [],
+        };
+      },
+    });
+
+    const networkAgent = new Agent({
+      id: 'orchestrator',
+      name: 'Orchestrator',
+      instructions: 'Coordinate research tasks',
+      model: routingMockModel,
+      agents: { 'research-agent': subAgent },
+      memory,
+    });
+
+    const stream = await networkAgent.network('Research and summarize', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: {
+        schema: resultSchema,
+      },
+    });
+
+    let agentExecutionSeen = false;
+
+    for await (const chunk of stream) {
+      if (chunk.type === 'agent-execution-end') {
+        agentExecutionSeen = true;
+      }
+    }
+
+    expect(agentExecutionSeen).toBe(true);
+
+    // Get structured result
+    const result = await stream.object;
+
+    expect(result).toBeDefined();
+    expect(result).toEqual(structuredResult);
+  });
+
+  it('should resolve object promise even if awaited before stream consumption', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: { schema: resultSchema },
+    });
+
+    // Start awaiting object before consuming stream
+    const objectPromise = stream.object;
+
+    // Now consume stream
+    for await (const _chunk of stream) {
+      // Process
+    }
+
+    // Object should resolve correctly
+    const result = await objectPromise;
+    expect(result).toBeDefined();
+    expect(result!.summary).toBe(structuredResult.summary);
+  });
+
+  it('should handle complex nested schemas', async () => {
+    const nestedSchema = z.object({
+      metadata: z.object({
+        version: z.string(),
+        timestamp: z.number(),
+      }),
+      results: z.array(
+        z.object({
+          id: z.string(),
+          data: z.object({
+            value: z.number(),
+            tags: z.array(z.string()),
+          }),
+        }),
+      ),
+      status: z.enum(['success', 'partial', 'failed']),
+    });
+
+    const nestedResult = {
+      metadata: { version: '1.0', timestamp: 1234567890 },
+      results: [
+        { id: 'r1', data: { value: 42, tags: ['important', 'urgent'] } },
+        { id: 'r2', data: { value: 17, tags: ['low'] } },
+      ],
+      status: 'success' as const,
+    };
+
+    const mockModel = createNetworkMockModel({ structuredResult: nestedResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test nested',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test nested schema', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: { schema: nestedSchema },
+    });
+
+    for await (const _chunk of stream) {
+      // Consume
+    }
+
+    const result = await stream.object;
+    expect(result).toBeDefined();
+    expect(result!.metadata.version).toBe('1.0');
+    expect(result!.results).toHaveLength(2);
+    expect(result!.results[0].data.tags).toContain('important');
+    expect(result!.status).toBe('success');
+  });
+
+  it('should handle schema with optional fields', async () => {
+    const optionalSchema = z.object({
+      required: z.string(),
+      optional: z.string().optional(),
+      withDefault: z.number().default(0),
+    });
+
+    const resultWithOptional = {
+      required: 'value',
+      // optional is omitted
+      withDefault: 5,
+    };
+
+    const mockModel = createNetworkMockModel({ structuredResult: resultWithOptional });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test optional',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test optional fields', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: { schema: optionalSchema },
+    });
+
+    for await (const _chunk of stream) {
+      // Consume
+    }
+
+    const result = await stream.object;
+    expect(result).toBeDefined();
+    expect(result!.required).toBe('value');
+    expect(result!.optional).toBeUndefined();
+    expect(result!.withDefault).toBe(5);
+  });
+
+  it('should resolve object as undefined when network completes without structuredOutput', async () => {
+    const mockModel = createNetworkMockModel({});
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      // No structuredOutput
+    });
+
+    for await (const _chunk of stream) {
+      // Consume
+    }
+
+    const result = await stream.object;
+    expect(result).toBeUndefined();
+  });
+
+  it('should accumulate usage across structured output generation', async () => {
+    const mockModel = createNetworkMockModel({ structuredResult });
+
+    const networkAgent = new Agent({
+      id: 'test-network',
+      name: 'Test Network',
+      instructions: 'Test',
+      model: mockModel,
+      memory,
+    });
+
+    const stream = await networkAgent.network('Test usage', {
+      requestContext,
+      completion: { scorers: [alwaysPassScorer as any] },
+      structuredOutput: { schema: resultSchema },
+    });
+
+    for await (const _chunk of stream) {
+      // Consume
+    }
+
+    const usage = await stream.usage;
+    // Usage should include tokens from both routing and structured output calls
+    expect(usage.totalTokens).toBeGreaterThan(0);
+  });
+});
